# Conflicts:
#	backend-services/CodaTrips/src/main/java/com/trip/coda/controllers/FlightsController.java
#	backend-services/CodaTrips/src/main/java/com/trip/coda/controllers/Hotelcontroller.java
#	backend-services/CodaTrips/src/main/java/com/trip/coda/repo/BookingInterface.java
#	backend-services/CodaTrips/src/main/resources/application.properties

+++ backend-services/CodaTrips/src/main/java/com/trip/coda/repo/BookingInterface.java
@@ -2,12 +2,12 @@ package com.trip.coda.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.trip.coda.models.BookingInput;
+
 import com.trip.coda.models.FlightBooking;
 
 
 
 
-public interface BookingInterface extends JpaRepository<FlightBooking,Long> {
+public interface BookingInterface extends JpaRepository<FlightBooking,Integer > {
 
 }
